refactor(web): extract MCP list cache into its own module
Move McpListCache type, mcpListCache state, getMcpListWorkingDir,
refreshMcpListCache and startMcpListChecker into src/web/mcp-list.ts.
Cache is module-private; routes read it via getMcpListCache().
Working-dir cleanup-on-exit hook moves with the module.

web.ts: 3980 -> 3844 lines.
